Add files via upload
This project is based on the problem of skin disorder . The EDA ( Task 1)of this project uncovered the factors which are affecting to the  different types of skin disorder. Task 2 of this project tells us that which factor is affecting to skin disorder most whereas Task 3 of this project contains model creation (Machine Learning Model).